i dont know here it is
Map shows up

+++ app/src/main/java/com/example/a403finalproject/WalkerList.java
@@ -121,8 +121,8 @@ public class WalkerList extends AppCompatActivity {
         queue = Volley.newRequestQueue(this);
 
         context = this;
-        double lat = (double)sharedPreferences.getFloat("lat",0);
-        double lon = (double)sharedPreferences.getFloat("lon", 0);
+        float lat = sharedPreferences.getFloat("lat",0);
+        float lon = sharedPreferences.getFloat("lon", 0);
 
         Log.d("lat and lon", lat + " " + lon);
